Change the (internal) parsing rules for header info.
The number of tokens can change depending on the PWscf executable, e.g.
appending a `svn rev. XXX` string after the version number. So instead of
parsing the date and time information using fixed indices, it's better to use
the immediately preceding words in the header (`on` for the date, `at` for the
time, `PWSCF` for the version number).

E.g.: "Program PWSCF v.6.1 (svn rev. 13591M) starts on 12Jul2017 at 10:17:52"
is still parsed correctly with the new rules.